LPS-12219 - When column is empty, dragging a new portlet in throws an error
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@60995 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-web/docroot/html/js/liferay/layout.js
@@ -121,7 +121,7 @@ AUI().add(
 										if (activeDropNode != lastActiveDropNode) {
 											var referencePortlet = Liferay.Layout.getLastPortletNode(activeDropNode);
 
-											if (referencePortlet.isStatic) {
+											if (referencePortlet && referencePortlet.isStatic) {
 												var options = Liferay.Layout.options;
 												var dropColumn = activeDropNode.one(options.dropContainer);
 												var foundReferencePortlet = Liferay.Layout.findReferencePortlet(dropColumn);
